[NFCI] Move `sycl::vec::convert` definition to `vector_convert.hpp`
1) It keeps all `convert`-related code in one place
2) It allows not to include that extra functionality if this particular
   method isn't used
3) Potentially, it moves all instances where `half`/`bfloat16` needs to
   be "complete" in `vector.hpp`. Can't really verify that because
   `generic_type_traits.hpp` includes definitions for both at the moment.
